Make code DRY
